Fixing AssertionsTest
